feat: Create createClaims method in TokenUtil file to contaim member information
rebase
